requirements.txt
astroid==1.6.2
autopep8==1.3.5
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.4
django-allauth==0.35.0
django-rest-auth==0.9.3
djangorestframework==3.8.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.0.7
Pillow==5.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pylint==1.8.3
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
wrapt==1.10.11
